fixed users

+++ app/controllers/groups.js
@@ -3,10 +3,35 @@ const db = require("../models");
 const Group = db.Groups;
 const User = db.Users;
 
+// create default group
+setTimeout(() => {
+  Group.findOrCreate({
+    where: { id: 1 },
+    defaults: {
+      name: "Group1",
+      type: "beg",
+      userId: 4,
+    },
+  }).then((res) => {
+    if (res[1]) {
+      console.log("Default group has been created successfully!");
+    } else {
+      console.log("Default group has been created!");
+    }
+  });
+}, 50);
+
 exports.get = (req, res) => {
+  const id = req.params?.id;
   let search = {};
   if (req.auth.role === "teacher") {
-    search.where = { userId: req.auth.id };
+    if (id) {
+      search.where = { userId: req.auth.id, id };
+    } else {
+      search.where = { userId: req.auth.id };
+    }
+  } else {
+    if (id) search.where = { id };
   }
   Group.findAll({ raw: true, ...search })
     .then((res1) => {
@@ -16,53 +41,64 @@ exports.get = (req, res) => {
 };
 
 exports.create = (req, res) => {
-  User.findByPk(req.auth.id)
-    .then((res1) => {
-      if (res1) {
-        User.findByPk(req.body.userId)
-          .then((res2) => {
-            if (res2 && res2?.role === "teacher") {
-              Group.findOne({ where: { type: req.body.type, name: req.body.name } })
-                .then((res3) => {
-                  if (!res3) {
-                    Group.create({
-                      name: req.body.name,
-                      type: req.body.type,
-                      userId: req.body.userId,
-                    })
-                      .then((res4) => {
-                        Group.findByPk(res4.id)
-                          .then((res5) => {
-                            return res.send(res5);
-                          })
-                          .catch((err5) => {
-                            return res.send(err5);
-                          });
+  if (req.auth.role !== "teacher") {
+    User.findByPk(req.auth.id)
+      .then((res1) => {
+        if (res1) {
+          User.findByPk(req.body.userId)
+            .then((res2) => {
+              if (res2 && res2?.role === "teacher") {
+                Group.findOne({ where: { type: req.body.type, name: req.body.name } })
+                  .then((res3) => {
+                    if (!res3) {
+                      Group.create({
+                        name: req.body.name,
+                        type: req.body.type,
+                        userId: req.body.userId,
                       })
-                      .catch((err4) => {
-                        return res.send(err4);
-                      });
-                  } else {
-                    return res.send("A group with this name and type exists!!!");
-                  }
-                })
-                .catch((err3) => res.send(err3));
-            } else {
-              return res.send("No such teacher or user exists!!!");
-            }
-          })
-          .catch((err2) => res.send(err2));
-      } else {
-        return res.sendStatus(401);
-      }
-    })
-    .catch((err) => res.send(err));
+                        .then((res4) => {
+                          return res.send("Group has been created successfully");
+                        })
+                        .catch((err4) => {
+                          return res.send(err4);
+                        });
+                    } else {
+                      return res.send("A group with this name and type exists!!!");
+                    }
+                  })
+                  .catch((err3) => res.send(err3));
+              } else {
+                return res.send("No such teacher or user exists!!!");
+              }
+            })
+            .catch((err2) => res.send(err2));
+        } else {
+          return res.sendStatus(401);
+        }
+      })
+      .catch((err) => res.send(err));
+  } else {
+    return res.sendStatus(401);
+  }
 };
 
 exports.update = (req, res) => {
-  User.findOne({ where: { id: req.body.userId } }).then((res1) => {
-    if (res1 && res1?.role === "teacher") {
+  if (req.auth.role !== "teacher") {
+    User.findByPk(req.auth.id).then(async (res1) => {
       let data = {};
+      if (req.body.userId) {
+        let message = "";
+        await User.findByPk(req.body.userId)
+          .then((res2) => {
+            if (res2 && res2?.role === "teacher") {
+              data.userId = req.body.userId;
+            } else {
+              message = "No such teacher or user exists!!!";
+            }
+          })
+          .catch((err2) => res.send(err2));
+        if (message.length !== 0) return res.send(message);
+      }
       if (req.body.name) {
         data.name = req.body.name;
       }
@@ -75,32 +111,34 @@ exports.update = (req, res) => {
       Group.update(data, { where: { id: req.params.id } })
         .then((res2) => {
           if (res2[0] !== 0) {
-            Group.findByPk(req.params.id)
-              .then((res3) => res.send(res3))
-              .catch((err3) => res.send(err3));
+            return res.send("Group updated successfully!");
           } else {
             return res.send("Not found");
           }
         })
         .catch((err1) => res.send(err1));
-    } else {
-      return res.send("No such teacher or user exists!!!");
-    }
-  });
+    });
+  } else {
+    return res.sendStatus(401);
+  }
 };
 
 exports.delete = (req, res) => {
-  Group.destroy({
-    where: { id: req.params.id },
-  })
-    .then((res1) => {
-      if (res1) {
-        res.send("Has been deleted!");
-      } else {
-        res.send("Not found");
-      }
+  if (req.auth.role !== "teacher") {
+    Group.destroy({
+      where: { id: req.params.id },
     })
-    .catch((err) => {
-      res.send(err);
-    });
+      .then((res1) => {
+        if (res1) {
+          res.send("Has been deleted!");
+        } else {
+          res.send("Not found");
+        }
+      })
+      .catch((err) => {
+        res.send(err);
+      });
+  } else {
+    return res.sendStatus(401);
+  }
 };
